+++ java/src/hype/extended/behavior/HTween.java
@@ -16,14 +16,16 @@ import hype.core.drawable.HDrawable;
 import hype.core.interfaces.HCallback;
 import hype.core.util.HConstants;
 import hype.core.util.HMath;
-import hype.core.util.HWarnings;
 import processing.core.PApplet;
 
 public class HTween extends HBehavior {
 	private HDrawable _target;
 	private HCallback _callback;
-	private float _s1,_s2,_s3, _e1,_e2,_e3;
-	private float _curr, _dcurr, _spring, _ease;
+	private float _s1, _s2, _s3;
+	private float _e1, _e2, _e3;
+	private float _curr1, _curr2, _curr3;
+	private float _origw, _origh;
+	private float _raw, _dRaw, _spring, _ease;
 	private int _property;
 	
 	public HTween() {
@@ -34,6 +36,10 @@ public class HTween extends HBehavior {
 	
 	public HTween target(HDrawable d) {
 		_target = d;
+		if(d != null) {
+			_origw = d.width();
+			_origh = d.height();
+		}
 		return this;
 	}
 	
@@ -50,16 +56,6 @@ public class HTween extends HBehavior {
 		return _callback;
 	}
 	
-	private void setScaleStart() {
-		if(_property!=HConstants.SCALE) return;
-		if(_target==null) {
-			HWarnings.warn(HWarnings.NULL_TARGET,"HTween.start()",null);
-		} else {
-			_s1 *= _target.width();
-			_s2 *= _target.height();
-		}
-	}
-	
 	public HTween start(float a) {
 		return start(a,a);
 	}
@@ -72,7 +68,6 @@ public class HTween extends HBehavior {
 		_s1 = a;
 		_s2 = b;
 		_s3 = c;
-		setScaleStart();
 		return this;
 	}
 	
@@ -92,16 +87,6 @@ public class HTween extends HBehavior {
 		return _s3;
 	}
 	
-	private void setScaleEnd() {
-		if(_property!=HConstants.SCALE) return;
-		if(_target==null) {
-			HWarnings.warn(HWarnings.NULL_TARGET,"HTween.end()",null);
-		} else {
-			_e1 *= _target.width();
-			_e2 *= _target.height();
-		}
-	}
-	
 	public HTween end(float a) {
 		return end(a,a);
 	}
@@ -114,7 +99,6 @@ public class HTween extends HBehavior {
 		_e1 = a;
 		_e2 = b;
 		_e3 = c;
-		setScaleEnd();
 		return this;
 	}
 	
@@ -161,20 +145,46 @@ public class HTween extends HBehavior {
 		return _property;
 	}
 	
+	public float nextRaw() {
+		_raw += (_dRaw) = (_dRaw*_spring + (1-_raw)*_ease);
+		float c = HMath.round512(_raw);
+		_curr1 = HMath.map(c,0,1,_s1,_e1);
+		_curr2 = HMath.map(c,0,1,_s2,_e2);
+		_curr3 = HMath.map(c,0,1,_s3,_e3);
+		return c;
+	}
+	
+	public float curr() {
+		return _curr1;
+	}
+	
+	public float curr1() {
+		return _curr1;
+	}
+	
+	public float curr2() {
+		return _curr2;
+	}
+	
+	public float curr3() {
+		return _curr3;
+	}
+	
 	@Override
 	public void runBehavior(PApplet app) {
 		if(_target==null) return;
 		
-		_curr += (_dcurr) = (_dcurr*_spring + (1-_curr)*_ease);
-		float c = HMath.round512(_curr);
-		float v1 = HMath.map(c,0,1,_s1,_e1);
-		float v2 = HMath.map(c,0,1,_s2,_e2);
-		float v3 = HMath.map(c,0,1,_s3,_e3);
+		float c = nextRaw();
+		float v1 = _curr1;
+		float v2 = _curr2;
+		float v3 = _curr3;
 		
 		switch(_property) {
 		case HConstants.WIDTH:		_target.width(v1); break;
 		case HConstants.HEIGHT:		_target.height(v1); break;
-		case HConstants.SCALE:		// same as SIZE
+		case HConstants.SCALE:
+			v1 *= _origw;
+			v2 *= _origh;
 		case HConstants.SIZE:		_target.size(v1,v2); break;
 		case HConstants.ALPHA:		_target.alpha(Math.round(v1)); break;
 		case HConstants.X:			_target.x(v1); break;
@@ -188,7 +198,7 @@ public class HTween extends HBehavior {
 		case HConstants.DLOC:		_target.move(v1,v1); break;
 		default: break;
 		}
-		if(c==1 && HMath.round512(_dcurr)==0) {
+		if(c==1 && HMath.round512(_dRaw)==0) {
 			unregister();
 			_callback.run(this);
 		}
@@ -201,7 +211,7 @@ public class HTween extends HBehavior {
 	
 	@Override
 	public HTween unregister() {
-		_curr = _dcurr = 0;
+		_raw = _dRaw = 0;
 		return (HTween) super.unregister();
 	}
 }